Move controller files and renamed package name.
