+++ internal/controller/login_controller.go
@@ -1,4 +1,4 @@
-package internal
+package controller
 
 import (
 	"net/http"
